add client folder
